+++ docs/versionhistory.rst
@@ -18,6 +18,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
   - ``anyio.receive_signals()`` -> ``anyio.open_signal_receiver()``
   - ``anyio.run_in_thread()`` -> ``anyio.run_sync_in_worker_thread()``
   - ``anyio.current_default_thread_limiter()`` -> ``anyio.current_default_worker_thread_limiter()``
+  - ``anyio.exceptions.ResourceBusyError`` -> ``anyio.exceptions.BusyResourceError``
 - **BACKWARDS INCOMPATIBLE** Added the ``acquire()`` and ``release()`` methods to the ``Lock``,
   ``Condition`` and ``Semaphore`` classes
 - **BACKWARDS INCOMPATIBLE** Removed the ``Event.clear()`` method. You must now replace the event